Use /api/morechildren for loading more comments
The /api/info endpoint did not give us the data of when a child
had further subcomments that were not included.

/api/morechildren is more sane.  However, it seems to put the
more_children type comments on the same hyrachy level as the
actual comment, not as a child.  To test, find a "load more" button
in an AskReddit that has > 20 children.  Click it and see the issue
with the placement of the "load more" buttons.