Improved Debian files and others
Major issues with the Debian control files have been fixed, though
there may still be a few unnoticed errors. The gitignore was also
updated to ignore Debian's build output, and the -C option was
added to scss in data/Makefile.am to prevent it from caching, and
thus producing an irreproducible build.