Refactor subentry to support viewing frontpage, fixes #56
Previously, we overzealously added '/r/' to things, which prevented the
frontpage from being viewed.

Additionally, we now always append a sort to subreddits when viewing
them.  E.g. `/r/linux` -> `/r/linux/hot`.  This is because the frontpage
can only be viewed with a sort (e.g. `/hot`) and not without (at `/`).