Fix icons in flatpak app
The custom icons were previously not displaying.  After some probing, it
appears that the flatpak version of GDK Pixbuf does not support loading
SVGs, so our SVG-only icons were not picked up.

To rectify this, I added some 24x24 sized pngs, which is the size they
are displayed at usually.